+++ frontend/src/components/chat/utils.ts
@@ -1,0 +1,21 @@
+import { format, isToday, isYesterday } from "date-fns";
+
+// Helper functions
+export const getInitials = (name: string) => {
+  return name
+    .split(" ")
+    .map((n) => n[0])
+    .join("")
+    .toUpperCase();
+};
+
+// Helper function to format dates for message groups
+export const getDateDisplay = (date: Date): string => {
+  if (isToday(date)) {
+    return "Today";
+  } else if (isYesterday(date)) {
+    return "Yesterday";
+  } else {
+    return format(date, "MMMM d, yyyy");
+  }
+};
